refactor: move the inventory endpoints out of the closure
Sixth group of #669. InventoryList, InventoryConfig, InventoryScan,
InventoryCryptoReport and InventoryAdopt move into
modules/api/resources_inventory.py.

Closure complexity 372 -> 335; resources.py 2796 -> 2618 lines.

This is the first group whose classes use the scope helpers. Those already
exist as module-level functions in resource_context, but they take the context
as their first argument, so rewriting every call site would have changed the
code being moved. The thin wrappers the closure prologue defines are
reproduced in the group factory instead, and the call sites move verbatim —
which is the property that makes an extraction reviewable at all.

That wiring is worth a test rather than a glance: a wrapper pointed at the
wrong function would either crash or, far worse, silently stop filtering and
hand a scoped caller the whole inventory. The new case builds the group with a
scoped caller and two records, calls InventoryList, and asserts the
out-of-scope one is absent. Verified by dropping the filter from the call: it
fails.

The manager-key guard added with the storage group validated this group's five
lookups without any new test being written for them, which is what it was for.
